refactor: extract duplicate engine utilities to shared packages
Extract `resolveFriendlyActorIds` to `packages/engine/src/processors/utils.ts`
and `parseAbilitySchoolMask` to `@wow-threat/shared/src/utils.ts`, removing
identical copies from party-detection.ts, tranquil-air-emulation.ts, wcl-input.ts,
and threat-aggregation.ts.
