+++ packages/engine/src/processors/tranquil-air-emulation.ts
@@ -10,6 +10,7 @@ import type {
 } from '../event-processors'
 import type { ActorId } from '../instance-refs'
 import { partyAssignmentsKey } from './party-detection'
+import { resolveFriendlyActorIds } from './utils'
 
 const TRANQUIL_AIR_TOTEM_SPELL_IDS = new Set<number>([25908])
 
@@ -52,17 +53,6 @@ function resolveFightFriendlyActorIds(
     ...(fight?.friendlyPlayers ?? []),
     ...(fight?.friendlyPets ?? []).map((pet) => pet.id),
   ])
-}
-
-function resolveFriendlyActorIds(
-  fightFriendlyActorIds: Set<ActorId>,
-  runtimeFriendlyActorIds: Set<ActorId> | undefined,
-): Set<ActorId> {
-  if (runtimeFriendlyActorIds && runtimeFriendlyActorIds.size > 0) {
-    return runtimeFriendlyActorIds
-  }
-
-  return fightFriendlyActorIds
 }
 
 function resolveOwnedPetIdsForOwners(
